feat: unify pdp client and server on common interface (#169)
The curx of the change here is aligning the PDP HTTP Client (previously
called curio client) and PDP service on a common interface. The
complexity of handling the various server request and response types
needed to continue supporting a curio http server has been pushed into
the PDP HTTP Client.