fix: join HTTP path to multiaddr (#180)
This joins a HTTP path to a multiaddr. The multiaddr is allowed to
already have an existing HTTP path. If it has an existing path it is
joined to the end of it, and if it does not have an existing HTTP path
then the "http-path" protocol is added to the multiaddr with the
specified value.