fix: do not smother HTTP handler errors (#197)
HTTP handler errors are being smothered by the Sentry reporter telemetry
when sentry is not configured. i.e. all piri instances except our own
that runs in AWS lambda.

This PR adds a middeware to the echo server which logs unhandled errors
(errors that are not `*echo.HTTPError`) using our usual logger.

It also alters the HTTP handler functions to be Echo compatible, using
an interface for handler context that is a subset of the `echo.Context`
interface. This allows the handlers to still be used in AWS lambda.

Since only our own AWS lambda Piri instance uses Sentry, I've removed
the wrapping (`telemetry.NewErrorReportingHandler(...)`) from the
handler constructors themselves, and moved it to the lambda entrypoints.

---------

Co-authored-by: frrist <forrest@storacha.network>