fix: specify range in location commitment (#229)
This PR adds the byte range of the blob to the location commitment. It
allows retrievals without having to trust the storage node to provide
the byte range.

i.e. it allows us to invoke `space/content/retrieve` with a byte range
so that we don't trust the storage node to report the bytes served on
it's own.

This is also consistent with location commitments issued by the legacy
upload service.