feat: fetch egress tracking proof on init (#243)
This PR depends on https://github.com/storacha/delegator/pull/11. It
adds config and flags for the egress tracking service and uses the new
delegator client method `RequestProofs` to fetch the delegation to
invoke `space/egress/track` on the egress tracking service, along with
the previous one to `claim/cache` on the indexer.