fix: concurrent IPNI advert publishing (#276)
This PR brings in the new IPNI publisher library that has a store with a
`Replace` method (see
https://github.com/storacha/go-libstoracha/pull/56).

The replace method allows the head to be updated safely when the backing
store is an S3 bucket that might be concurrently written to by multiple
lambdas.

The replace method makes use of `If-Match` HTTP header in order to only
write the value if the current value has the same `Etag`.

We also try really hard to write an advert - this PR adds retries to the
advert publish call that allows the conditional write to fail up to 5
times.