feat: authorized retrieval for replica transfer (#286)
This PR impements authorized retrieval for replica transfer.

In order to authorize retrieval it first makes a `access/grant`
invocation to the replication source requesting a `blob/retrieve`
delegation. It attaches the `blob/replica/allocate` invocation as
evidence to the source that the request is legitimate.

It then uses the delegation in a `blob/retrieve` invocation to the
source, allowing retrieval of the data.