fix: only track egress caused by space/content/retrieve invocations (#303)
The retrieve server is configured with a receipt logger so that we can
collect `space/content/retrieve` receipts. However, the receipt logger
is configured for the whole server, which also serves `blob/retrieve`
invocations. This PR filters receipts so that only
`space/content/retrieve` ones are collected for egress tracking.