fix: increase allowed transfer time (#318)
This PR increases the transfer task expiration from 30 seconds to 1
hour. It allows for queuing delays and retries. This time is effectively
the time we give ourselves to transfer the blob and send a
`ucan/conclude` invocation to the upload service.

Aims to resolve:

```
'Unauthorized: Claim {"can":"blob/replica/transfer"} is not authorized\n' +
          '  - Proof bafyreig3wen5tsemfffi6u4tfou4y7v2mffajuyujgc53vpvoykbzdf6pe has expired on Thu Oct 30 2025 16:17:30 GMT+0000 (Coordinated Universal Time)\n'
```

...which is the upload service validating the transfer invocation in the
`ucan/conclude` invocation.