docs: seperate guide for forge production (#373)
This is a copy of the getting start guide docs with edits to make them
relevant for the production network.

I'm not in love with this. There's a lot of repetition, but because the
docs all link to each other they all kinda need to be copied otherwise
you fall out of the production specific docs.