feat(service): only calculate piece cids once (#390)
# Goals

move writing the Piece:MH DB record to commP calc. moreover, make sure
that only one calc is in process for a given multihash at any time

# Implementation

- key library is singleflight which I guess is golang stdlib but I
didn't know about it. basically allows you to share a function call
between go routines so it only gets called for a given set of params
once at a time.
- critically, moved writing the PDPPieceCIDToMH record to this function,
so that calculation is only ever called once for each multihash (and
this will fortunately cause future calls to the piece resolver to
succeed)
- wrote a test cause it's a bit complicated, man this whole section is
woefully undertested